Added user input for how many of each shape can be selected
Used a series of Arrays and while loops to allow the user to input their desired amounts of each shape and then fill in the information accordingly. This simplifies things and stops the user from having to input values of 0 for each value used in calculations if the shape is not included. Application now caters to all users rather than one specific business. Will be creating a front end for this application in a new project. will be either windows form application or web application.